requirements.txt
flask
Pillow
cv2
werkzeug
numpy